requirements.txt
discord.py
python-dotenv
github
pymongo
chat_exporter